Add button Fixes
The add button rotation was fixed. It should no longer get out of sync if the user clicks the button rapidly. There is now a popup that appears when you click the "Manually Add New Game" button. For the game system ComboBox, if you select "Add New System" a TextField will appear to the right of the box and you can enter a system manually. If the field is visible and you choose a system from the list the TexField will disappear. Finally when you click the "Add" button the information you inputted will display in console. Currently the information doesn't save and there is no field checking; meaning if a field is empty there will be no error. However, there is a check in place in the "hoursField"; this field will only accept numeric values.

Side Note: Any links I used is preceded by "<LINK>". The "link tag" can be removed. I only put them in so the project manager could more easily find my sources.